[Fix](BE) : 회원권 소지 시 0원으로 정산하게 수정, 정산 금액 overflow 문제 해결

See merge request s11-webmobile3-sub2/S11P12E101!80

+++ BE/mvp/src/main/java/com/mvp/logger/dto/VehicleLogDTO.java
@@ -12,7 +12,7 @@ public class VehicleLogDTO {
     private String licensePlate;
     private LocalDateTime entranceTime;
     private LocalDateTime exitTime;
-    private Integer fee;
+    private Long fee;
     private byte[] image;
     private Integer type;
 }
